+++ scripts/setup-server-front.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-gem-install-if-needed sass
-npm-install-if-needed gulp bower
-
 cat > /home/${username}/taiga-front/app/config/main.coffee <<EOF
 config = {
     host: "${hostname}"
@@ -23,10 +20,22 @@ config = {
 angular.module("taigaLocalConfig", []).value("localconfig", config)
 EOF
 
-sudo rm -rf /home/#{username}/tmp
+pushd ~
 
-pushd ~/taiga-front
-npm install
-bower install
-gulp deploy
+if [ ! -e ~/taiga-front ]; then
+    git clone https://github.com/taigaio/taiga-front.git taiga-front
+
+    gem-install-if-needed sass
+    npm-install-if-needed gulp bower
+
+    sudo rm -rf /home/#{username}/tmp
+
+    pushd ~/taiga-front
+    npm install
+    bower install
+    gulp deploy
+    popd
+fi
+
 popd
+
